package: usbmount: add config for "sync" mount option
The "sync" option may not be a good choice to use with flash drives, as
it forces a greater amount of writing operating on the drive. This makes
the writing speed considerably lower and also leads to a faster wear out
of the disk.

Let's remove "sync" option as default, and add a compile config to control it.

Change-Id: Ifca8de4bc97380bacec921d6967814e7c82b72b5
Signed-off-by: zain wang <wzz@rock-chips.com>